# lib/api-zod/src/generated/types/boardMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BoardMember {
  id: number;
  name: string;
  role: string;
  bio: string;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}