# lib/api-zod/src/generated/types/boosterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BoosterInput {
  /** @minLength 1 */
  title: string;
  description: string;
  type: string;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
}